python_ai/requirements.txt
# requirements.txt
# Stock Prediction System - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
lightgbm==4.1.0
scikit-learn==1.3.2

# Stock Data
yfinance==0.2.32

# NLP & Sentiment Analysis
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# LLM Integration
openai==1.3.7

# Scheduler
APScheduler==3.10.4

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional (for enhanced features)
beautifulsoup4==4.12.2
lxml==4.9.3